push last code of java to start mongo

+++ src/main/java/com/example/examplequerydslspringdatajpamaven/photo/DecodePhoto.java
@@ -5,11 +5,64 @@ import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
+import java.util.ArrayList;
 import java.util.Base64;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.Random;
 
 public class DecodePhoto {
 	
+	
+	public Boolean deleteIcon(String icon,String modleType) {
+		String path ="";
+		if(modleType.equals("icon")) {
+			path ="/var/www/html/sareb_photo/icons/"+icon;
+		}
+	   
+		File file = new File(path);
+		if (file.isFile()) {
+			file.delete();
+		}
+		return true;
+        
+	}
+	public List<Map> getAllIcons() {
+		String path ="/var/www/html/sareb_photo/icons/default/";
+		File folder = new File(path);
+
+		File[] listOfFiles = folder.listFiles();
+		List<Map> dataMap=new ArrayList<>();
+		
+		for (int i = 0; i < listOfFiles.length; i++) {
+			
+		    Map<String, String> list =new HashMap<String, String>();
+			if (listOfFiles[i].isFile()) {
+				list.put("icon",listOfFiles[i].getName());
+			}
+			dataMap.add(list);
+		}
+		
+		return dataMap;
+		
+        
+	}
+	public Boolean checkIconDefault(String icon,String modleType) {
+		String path ="";
+		if(modleType.equals("icon")) {
+			path ="/var/www/html/sareb_photo/icons/default/"+icon;
+		}
+	   
+		File file = new File(path);
+		if (file.isFile()) {
+			return true;
+		}
+		else {
+			return false;
+		}
+        
+	}
 	
 	public Boolean deletePhoto(String photo,String modleType) {
 		String path ="";
@@ -23,8 +76,10 @@ public class DecodePhoto {
 			path ="/var/www/html/sareb_photo/vehicle_photos/"+photo;
 		}
 		File file = new File(path);
-		
-		return file.delete();
+		if (file.isFile()) {
+			file.delete();
+		}		
+		return true;
 	}
 	public String Base64_Image(String photo,String modleType) {
 		
@@ -59,8 +114,10 @@ public class DecodePhoto {
 		if(modleType.equals("vehicle")) {
 			path ="/var/www/html/sareb_photo/vehicle_photos/"+fileName;
 		}
+		if(modleType.equals("icon")) {
+			path ="/var/www/html/sareb_photo/icons/"+fileName;
+		}
 		
-
 		int pos=photo.indexOf(",");
 		byte[] data=Base64.getDecoder().decode(photo.substring(pos+1));
